feat(payments): 결제 상품 목록 리사이클러뷰 설정
- 결제 어댑터 설정 OrderAdapter.kt
- 리사이클러뷰 설정 및 어댑터 연결 setupRecyclerView()
- 결제 완료 응답값 1차원 리스트로 변환 convertToOrderItems()
- 상품 아이템 개별 가격에서 총 가격으로 변경 price -> totalPrice
- 홈 이동 버튼 설정 (finish)

+++ app/src/main/java/com/nuda/nudaclient/data/remote/dto/shopping/ShoppingCreateOrderResponse.kt
@@ -2,7 +2,7 @@ package com.nuda.nudaclient.data.remote.dto.shopping
 
 data class ShoppingCreateOrderResponse(
     val orderId: Int,
-    val orderNum: Int,
+    val orderNum: Long,
     val status: String,
     val totalAmount: Int,
     val brands: List<Brand>
