+++ app/src/main/java/com/nuda/nudaclient/data/remote/dto/shopping/ShoppingCreatePaymentsResponse.kt
@@ -3,7 +3,7 @@ package com.nuda.nudaclient.data.remote.dto.shopping
 data class ShoppingCreatePaymentsResponse(
     val paymentId: Int,
     val orderId: Int,
-    val orderNum: Int,
+    val orderNum: Long,
     val amount: Int,
     val status: String,
     val paymentKey: String,
